// src/colab/client/generated/schema-v1.d.ts
export interface paths {
    "/v1/operations": {
        parameters: {
            query?: {
                /** @description Data format for response. */
                $alt?: components["parameters"]["alt"];
                /** @description JSONP */
                $callback?: components["parameters"]["callback"];
                /** @description Returns response with indentations and line breaks. */
                $prettyPrint?: components["parameters"]["prettyPrint"];
                /** @description V1 error format. */
                "$.xgafv"?: components["parameters"]["_.xgafv"];
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * @description Lists operations that match the specified filter in the request. If the
         *     server doesn't support this method, it returns `UNIMPLEMENTED`.
         */
        get: operations["ListOperations"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/operations/{operationsId}:cancel": {
        parameters: {
            query?: {
                /** @description Data format for response. */
                $alt?: components["parameters"]["alt"];
                /** @description JSONP */
                $callback?: components["parameters"]["callback"];
                /** @description Returns response with indentations and line breaks. */
                $prettyPrint?: components["parameters"]["prettyPrint"];
                /** @description V1 error format. */
                "$.xgafv"?: components["parameters"]["_.xgafv"];
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /**
         * @description Starts asynchronous cancellation on a long-running operation.  The server
         *     makes a best effort to cancel the operation, but success is not
         *     guaranteed.  If the server doesn't support this method, it returns
         *     `google.rpc.Code.UNIMPLEMENTED`.  Clients can use
         *     Operations.GetOperation or
         *     other methods to check whether the cancellation succeeded or whether the
         *     operation completed despite cancellation. On successful cancellation,
         *     the operation is not deleted; instead, it becomes an operation with
         *     an Operation.error value with a google.rpc.Status.code of `1`,
         *     corresponding to `Code.CANCELLED`.
         */
        post: operations["CancelOperation"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/operations/{operationsId}": {
        parameters: {
            query?: {
                /** @description Data format for response. */
                $alt?: components["parameters"]["alt"];
                /** @description JSONP */
                $callback?: components["parameters"]["callback"];
                /** @description Returns response with indentations and line breaks. */
                $prettyPrint?: components["parameters"]["prettyPrint"];
                /** @description V1 error format. */
                "$.xgafv"?: components["parameters"]["_.xgafv"];
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * @description Gets the latest state of a long-running operation.  Clients can use this
         *     method to poll the operation result at intervals as recommended by the API
         *     service.
         */
        get: operations["GetOperation"];
        put?: never;
        post?: never;
        /**
         * @description Deletes a long-running operation. This method indicates that the client is
         *     no longer interested in the operation result. It does not cancel the
         *     operation. If the server doesn't support this method, it returns
         *     `google.rpc.Code.UNIMPLEMENTED`.
         */
        delete: operations["DeleteOperation"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        Operation: components["schemas"]["BaseOperation"] & {
            /**
             * @description Service-specific metadata associated with the operation.  It typically
             *     contains progress information and common metadata such as create time.
             *     Some services might not provide such metadata.  Any method that returns a
             *     long-running operation should document the metadata type, if any.
             */
            metadata?: {
                [key: string]: unknown;
            };
            /**
             * @description The normal, successful response of the operation.  If the original
             *     method returns no data on success, such as `Delete`, the response is
             *     `google.protobuf.Empty`.  If the original method is standard
             *     `Get`/`Create`/`Update`, the response should be the resource.  For other
             *     methods, the response should have the type `XxxResponse`, where `Xxx`
             *     is the original method name.  For example, if the original method name
             *     is `TakeSnapshot()`, the inferred response type is
             *     `TakeSnapshotResponse`.
             */
            response?: {
                [key: string]: unknown;
            };
        };
// ...
    };
// ...
}