// src/colab/client/generated/schema-v1beta.d.ts
export interface paths {
    "/v1beta/subscription": {
        parameters: {
            query?: {
                /** @description Data format for response. */
                $alt?: components["parameters"]["alt"];
                /** @description JSONP */
                $callback?: components["parameters"]["callback"];
                /** @description Returns response with indentations and line breaks. */
                $prettyPrint?: components["parameters"]["prettyPrint"];
                /** @description V1 error format. */
                "$.xgafv"?: components["parameters"]["_.xgafv"];
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description Gets the subscription of the requesting user. */
        get: operations["GetSubscription"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1beta/runtimes": {
        parameters: {
            query?: {
                /** @description Data format for response. */
                $alt?: components["parameters"]["alt"];
                /** @description JSONP */
                $callback?: components["parameters"]["callback"];
                /** @description Returns response with indentations and line breaks. */
                $prettyPrint?: components["parameters"]["prettyPrint"];
                /** @description V1 error format. */
                "$.xgafv"?: components["parameters"]["_.xgafv"];
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description Lists Colab runtimes assigned to the requesting user. */
        get: operations["ListRuntimes"];
        put?: never;
        /** @description Creates a Colab runtime assignment. */
        post: operations["CreateRuntime"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1beta/runtimespecs": {
        parameters: {
            query?: {
                /** @description Data format for response. */
                $alt?: components["parameters"]["alt"];
                /** @description JSONP */
                $callback?: components["parameters"]["callback"];
                /** @description Returns response with indentations and line breaks. */
                $prettyPrint?: components["parameters"]["prettyPrint"];
                /** @description V1 error format. */
                "$.xgafv"?: components["parameters"]["_.xgafv"];
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description Lists Colab runtime specs available to the requesting user. */
        get: operations["ListRuntimeSpecs"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1beta/runtimes/{runtime}": {
        parameters: {
            query?: {
                /** @description Data format for response. */
                $alt?: components["parameters"]["alt"];
                /** @description JSONP */
                $callback?: components["parameters"]["callback"];
                /** @description Returns response with indentations and line breaks. */
                $prettyPrint?: components["parameters"]["prettyPrint"];
                /** @description V1 error format. */
                "$.xgafv"?: components["parameters"]["_.xgafv"];
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description Gets a Colab runtime assigned to the requesting user. */
        get: operations["GetRuntime"];
        put?: never;
        post?: never;
        /** @description Deletes a Colab runtime assignment. */
        delete: operations["DeleteRuntime"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        Operation: components["schemas"]["BaseOperation"] & {
            /**
             * @description Service-specific metadata associated with the operation.  It typically
             *     contains progress information and common metadata such as create time.
             *     Some services might not provide such metadata.  Any method that returns a
             *     long-running operation should document the metadata type, if any.
             */
            metadata?: {
                [key: string]: unknown;
            };
            /**
             * @description The normal, successful response of the operation.  If the original
             *     method returns no data on success, such as `Delete`, the response is
             *     `google.protobuf.Empty`.  If the original method is standard
             *     `Get`/`Create`/`Update`, the response should be the resource.  For other
             *     methods, the response should have the type `XxxResponse`, where `Xxx`
             *     is the original method name.  For example, if the original method name
             *     is `TakeSnapshot()`, the inferred response type is
             *     `TakeSnapshotResponse`.
             */
            response?: {
                [key: string]: unknown;
            };
        };
        /** @description A specification of a runtime. */
        RuntimeSpec: {
            /** @description Required. The key of the runtime spec. */
            key: components["schemas"]["Key"];
            /** @description Output only. Whether the requesting user is eligible for the runtime spec. */
            readonly eligible?: boolean;
        };
        /** @description Metadata for `CreateRuntime` long-running operation. */
        CreateRuntimeMetadata: {
            /**
             * @description The runtime resource being created, in the format of
             *     `runtimes/{runtime_id}`.
             */
            runtime?: string;
        };
        /** @description Key fields that uniquely identify a runtime spec. */
        Key: {
            /** @description Required. Runtime variant. */
            variant: components["schemas"]["Variant"];
            /**
             * @description Required. Runtime accelerator model name (e.g., "T4", "L4", "A100").
             *     For CPU runtime specs, this field will be "NONE".
             */
            accelerator: string;
            /** @description Required. Runtime shape. */
            shape: components["schemas"]["Shape"];
        };
// ...
        Subscription: {
            /** @description Identifier. Name of the subscription. This will always be "subscription". */
            name?: string;
            /** @description User's Colab subscription tier. */
            tier?: components["schemas"]["SubscriptionTier"];
        };
// ...
        CreateRuntimeOperation: components["schemas"]["BaseOperation"] & {
            metadata?: components["schemas"]["CreateRuntimeMetadata"];
            response?: components["schemas"]["Runtime"];
        };
// ...
        Runtime: {
            /**
             * @description Identifier. The name of the runtime in the format `runtimes/{runtime_id}`, where
             *     `runtime_id` is either the value supplied via
             *     `CreateRuntimeRequest.runtime_id` or, if none was supplied, a random UUID
             *     generated by the backend.
             */
            name?: string;
            /** @description Required. The specification of the runtime. */
            runtimeSpec: components["schemas"]["Key"];
            /** @description Output only. The connection info used to authenticate to the runtime. */
            readonly connectionInfo?: components["schemas"]["ConnectionInfo"];
        };
// ...
    };
// ...
}